+++ app/src/main/java/de/hamedtanha/servertoolkit/core/connection/domain/model/ConnectionTargetInvalidReason.kt
@@ -1,0 +1,7 @@
+package de.hamedtanha.servertoolkit.core.connection.domain.model
+
+enum class ConnectionTargetInvalidReason {
+    MissingHost,
+    InvalidPort,
+    MissingUsername,
+}