+++ app/src/main/java/de/hamedtanha/servertoolkit/core/connection/domain/resolver/ConnectionTargetResolver.kt
@@ -1,0 +1,8 @@
+package de.hamedtanha.servertoolkit.core.connection.domain.resolver
+
+import de.hamedtanha.servertoolkit.core.connection.domain.model.ConnectionTargetResolution
+
+interface ConnectionTargetResolver {
+
+    suspend fun resolve(serverId: String): ConnectionTargetResolution
+}